Add basic screenshotting ability and site statistics output
Define and implement a WebRecorder class with the capability to
save screenshots of sites and output both the site's IP address
and the time for the site to load. Add a main routine with the
ability to parse and except both positional and optional arguments
upon program execution.

The class provides a basic framework from which to begin building
up a backend, and the main routine creates a structure that can
be used to sustainably runtime options.