Add Geolocation Spoofing Ability
Switch over to using the Chrome browser to visit a given url. This
allows for the utilization of the command driver.execute_cdp_cmd()
in order to change the location settings in the started browser
to a set of coordinates. In this version, these coordinates are
hard-coded to be (52.520007, 13.404954) -- Berlin, Germany.