Update documentation, combine startup scripts, fix image location
Add information about the implementation and purpose of Project
Theia to better inform others about the application. Combine
the bash scripts front-end.sh and back-end.sh into a single script
start.sh which runs the commands "npm start" and "node app.js" in
separate processes. This streamlines the startup process for the
user.

Fix a bug in web_recorder.py, which set the default save location
to an invalid file path, causing the script to throw an exception.
Doing this allows the application to function without any
modification.